requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.1.0
aiosqlite==0.19.0
p115client==0.0.8.4.2
loguru==0.7.2
pyyaml==6.0.3
pydantic==2.12.5
pydantic-settings==2.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
